batch parse update
much progress on the batch parse. need to modify the operators which were in most cases outputting blank data due to the randomness of the location of the objects within each line's json structure. (searching for "','" string literal would pull the first one on the line, regardless of whether it was before or after the object name string literal)